requirements.txt
Flask>=3.0,<4.0
boto3>=1.34,<2.0
SQLAlchemy>=2.0,<3.0
kafka-python>=2.0,<3.0
redis>=5.0,<6.0
psycopg[binary]>=3.1,<4.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<24.0
pytest>=8.0,<9.0
